requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.30.0
tensorflow>=2.13.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
librosa>=0.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
umap-learn>=0.5.3

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0 